osbase/bootparam: make test external
The bootparam package was the only package with this edge case:

1. It has both an internal and an external test.
2. The external test depends on a separate package, which depends on the
   package under test.

The core of rules_go has a 225 line function to handle this edge case:
https://github.com/bazel-contrib/rules_go/blob/v0.55.1/go/private/rules/test.bzl#L519

However, the edge case is not handled by the gopackagesdriver.
Additionally, the separate package (bootparam/ref) uses cgo, which
causes additional problems. This means that the Go VS Code extension
always reports 8 warnings in the bootparam and ref packages.

Instead of fixing this edge case in gopackagesdriver, it's easier to
make bootparam package not have this special case by making the internal
test external. This removes all the warnings.

Change-Id: Ia0aa4d2b62cdeaaec9983f3ecf4d229ea49358b2
Reviewed-on: https://review.monogon.dev/c/monogon/+/4387
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>
Tested-by: Jenkins CI