treewide: update etcd to 3.6
This is a fairly large update, containing 4 years of upstream
development. A lot of code has been moved around, requiring a rebase of
all patches and ajustments in code using it.

Upstream requires that upgrades first go through etcd 3.5.20, which is
done in a CL below this one. Other than that upgrades are possible
through normal Monogon node updates, but downgrades (while implemented
for 3.6.0) are still not supported in Monogon until further work is
done.

There are significant issues in etcd synchronization between the
"outside" manager (embed or standalone) and the core (EtcdServer), one
of which affects removal of the ConsensusMember role, causing a panic
due to the client listener not being closed when the server shuts down
on its own. This is triggered by the autopromoter hitting an endpoint on
a shut-down etcd, which accesses a nil backend. This issue existed
before, a full fix will likely involve either significantly rewriting or
getting rid of the embed package, so this just fixes the panic-causing
code path.

Change-Id: If5932a7428a262fde406a5bb652a40d211301734
Reviewed-on: https://review.monogon.dev/c/monogon/+/4394
Tested-by: Jenkins CI
Reviewed-by: Tim Windelschmidt <tim@monogon.tech>