gomod: update etcd to 3.5.21
This is required for any in-place upgrades to etcd 3.6 and higher as
earlier 3.5 releases have a bug [1] relating to learner metadata which
all Monogon nodes would trigger. Any ConsensusMember nodes first need to
be upgraded to a build containing this change but not the 3.6 upgrade
before being able to upgrade to a 3.6-containg version.

[1] https://github.com/etcd-io/etcd/issues/19557

Change-Id: I689b3456daf0fa28c38a8cc22e33cb4b6032ad49
Reviewed-on: https://review.monogon.dev/c/monogon/+/4393
Reviewed-by: Tim Windelschmidt <tim@monogon.tech>
Tested-by: Jenkins CI