third_party/linux: make build reproducible
Linux includes a timestamp, username and hostname in /proc/version and
other places. We need to set these to a fixed value to make the build
reproducible. For the timestamp, we can use the modified date of any
file in the source archive; all files in the archive have the same
timestamp.

See Linux Documentation/kbuild/reproducible-builds.rst

Change-Id: If50bb5a8ee73e5bc37e95a5d38e7ee87d85518a5
Reviewed-on: https://review.monogon.dev/c/monogon/+/4417
Tested-by: Jenkins CI
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>