third_party/edk2: use busybox toolchain
This add the busybox toolchain as we need it as soon as the tools are
not in PATH. Currently its fine as we have a FHSEnv on NixOS and
coreutils in our Jenkins CI. As soon as we remove the FHSEnv this is
actually required. It's done by replacing the strace toolchain which
was added by accident when debugging the edk2 build.

Change-Id: I081099d93ea3625a00d216569f522154ba3faabe
Reviewed-on: https://review.monogon.dev/c/monogon/+/4416
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>
Tested-by: Jenkins CI