osbase/build/mkpayload: make section order reproducible
Iteration order of Go maps is randomized, which resulted in sections
being added to the EFI file in random order. Now, the sections are
stored in a slice instead of a map, such that they are added to the file
in the order they are defined. This fixes a reproducibility bug.

Change-Id: I2a4c775a6b16e4f8a00c6b195dd169d874213d47
Reviewed-on: https://review.monogon.dev/c/monogon/+/4441
Tested-by: Jenkins CI
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>