metropolis/node: generate OCI index
This adds a multi-architecture OCI index target for metropolis. mkoci is
extended to add the platform to the index.json. Product info is extended
to add the PlatformOS field, which mkoci puts into the platform.

The OCI spec requires the os field of the platform to be set. We don't
need this field so we could simply put "unknown" there, but I think
"uefi" is nicer. Registry UIs display the os and architecture fields.

Change-Id: I24acf3c01201b50238abd30cfd86da2eda43c1d9
Reviewed-on: https://review.monogon.dev/c/monogon/+/4477
Tested-by: Jenkins CI
Reviewed-by: Tim Windelschmidt <tim@monogon.tech>