m/n/core: simplify network overlay
During the initial refactor into the role server the node external IP
(which was then used as the underlay target) got moved into the
published prefixes. This never did anything as the summary route into
clusternet does not contain it. Thus drop this and simplify the code.

Also remove the extra loopback IP introduced in 011dce6623 as it never
worked because control plane nodes do not even have a prefix to pick
such an IP from.

While at it, rename some variables to make them more accurate with how
this is eventually going to look. 

Change-Id: I6a6a6964bd45d088c37f746d1c9514246d624093
Reviewed-on: https://review.monogon.dev/c/monogon/+/4524
Tested-by: Jenkins CI
Reviewed-by: Tim Windelschmidt <tim@monogon.tech>