build/http_archive_deb: introduce
This adds a new repo rule which is essentially an extended version of
http_archive specifically for extracting Debian packages into repos.
http_archive can extract these, but only the outer AR layer which
leaves you with a data.tar.xyz and control.tar.xyz which is not what you
want most of the time.

This rule does that first step, but then figures out the respective file
extensions and extracts both of the tarballs, with the control tarball
going into a "debian" subdirectory. This allows grabbing files from the
archive directly.

Change-Id: I6a6a6964e6202526ee9093488f9c10e709f19149
Reviewed-on: https://review.monogon.dev/c/monogon/+/4649
Reviewed-by: Jan Schär <jan@monogon.tech>
Tested-by: Jenkins CI