m/n/k/plugins/kvmdevice: allow everyone to access /dev/kvm
This changes the permissions of /dev/kvm to allow everyone access.
Previously, only the owner, which is root, had access.

This allows containers which don't run as root to use KVM. It also makes
KVM accessible in user namespace containers, where the owner is mapped
to nobody, such that access is only possible if "other" has permissions.

Change-Id: Ie438d2ab40468a6d3002c3890b75b9c43188198d
Reviewed-on: https://review.monogon.dev/c/monogon/+/4654
Tested-by: Jenkins CI
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>