m/node/kubernetes: fix mount flags in userns containers
This change reverts commit ff7452b58613, and extends the persistent
volume test, which validates mount flags, to also run in a user
namespace.

The change to set noexec on the host was made because runc replaces the
flags inside the container, so they can be set restrictively on the
host. As we now found out, this doesn't work when the container has a
user namespace. The reason is that Linux locks the noexec and similar
flags when creating the mount namespace in the user namespace. At the
point where runc tries to replace the flags, they are already locked.

The solution for now is to not set the flags on the host. The ideal
solution would be for runc to apply the flags before they are locked,
and then lock them so the container can't clear them. But implementing
this is much more effort, and currently we can't set mount flags through
the Kubernetes API anyway.

See: https://github.com/monogon/monogon/issues/361
Change-Id: I490280c3c15397f288c2fab163857fa67ae5a721
Reviewed-on: https://review.monogon.dev/c/monogon/+/4657
Tested-by: Leopold Schabel <leo@monogon.tech>
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>